Node can decide whether it can be deleted

+++ egui_node_graph/src/editor_ui.rs
@@ -771,9 +771,16 @@ where
         // --- Interaction ---
 
         // Titlebar buttons
-        if Self::close_button(ui, outer_rect).clicked() {
-            responses.push(NodeResponse::DeleteNodeUi(self.node_id));
-        };
+        let can_delete = self.graph.nodes[self.node_id].user_data.can_delete(
+            self.node_id,
+            self.graph,
+            user_state,
+        );
+        if can_delete {
+            if Self::close_button(ui, outer_rect).clicked() {
+                responses.push(NodeResponse::DeleteNodeUi(self.node_id));
+            };
+        }
 
         let window_response = ui.interact(
             outer_rect,
